Remove unnecessary pytest helper files
