+++ tests/fields/test_constructors.py
@@ -6,8 +6,7 @@ import numpy as np
 
 import galois
 
-from ..helper import randint
-from .helper import valid_dtype, invalid_dtype
+from .conftest import randint, valid_dtype, invalid_dtype
 
 
 RANDOM_REPRODUCIBLE = {
